+++ src/app/admin/courselist/courselist.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { Router,ActivatedRoute} from '@angular/router';
 import { CourseServiceService } from 'src/app/course-service.service';
 
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from 'src/app/confirm-dialog/confirm-dialog.component';
+
 @Component({
   selector: 'app-courselist',
   templateUrl: './courselist.component.html',
@@ -34,7 +37,7 @@ export class CourselistComponent implements OnInit {
   }];
  
 t:any;
-  constructor(private _courseService:CourseServiceService,private _router:Router,private route:ActivatedRoute) { }
+  constructor(private _courseService:CourseServiceService,private _router:Router,private route:ActivatedRoute,private dialog: MatDialog) { }
 
   ngOnInit(): void {
 
@@ -52,14 +55,31 @@ this.t=type;
 
 
    deleteCourse(name:any){
-      
+      //--------------------dialog-------------
+      const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+        data: {
+          title: 'Confirm Delete Course',
+         
+        }
+      });
+      confirmDialog.afterClosed().subscribe(result => {
+        if (result === true) {
+         
+          this._courseService.deleteCourse(name)
+          .subscribe((data) => {
+            this.data = this.data.filter(b => b !== name);
+            //this.books = JSON.parse(JSON.stringify(data));
+            window. location. reload();
+            
+          });
+
+        }
+      });
+
+
+      //------------------------------------------
    
-  this._courseService.deleteCourse(name)
-    .subscribe((data) => {
-      this.data = this.data.filter(b => b !== name);
-      //this.books = JSON.parse(JSON.stringify(data));
-      
-    });
+  
 
 }
